reorganize services package
