+++ app/src/main/java/fho/kdvs/global/SharedViewModel.kt
@@ -45,6 +45,8 @@ import fho.kdvs.news.NewsRepository
 import fho.kdvs.schedule.QuarterRepository
 import fho.kdvs.schedule.QuarterYear
 import fho.kdvs.services.*
+import fho.kdvs.services.notification.CustomAction
+import fho.kdvs.services.notification.PlaybackType
 import fho.kdvs.show.ShowRepository
 import fho.kdvs.staff.StaffRepository
 import fho.kdvs.subscription.SubscriptionRepository
